moving spec files to __tests__
